feat: umap params as separate component
