+++ frontend/src/discovery/EmbedView.svelte
@@ -1,11 +1,16 @@
 <script lang="ts">
 	import Scatter from "./scatter/Scatter.svelte";
 	import Button from "@smui/button";
-	import { colorSpec } from "../stores";
-	import type { ScatterRowsWithIds } from "./scatter/scatter";
+	import UMAPParams from "./UMAPParams.svelte";
+
+	import { colorSpec, filteredTable, settings } from "../stores";
+	import { columnHash } from "../util/util";
 	import { onMount } from "svelte";
 	import { model } from "../stores";
 	import request from "../util/request";
+
+	import type ColumnTable from "arquero/dist/types/table/column-table";
+	import type { ScatterRowsWithIds } from "./scatter/scatter";
 
 	type point2D = [number, number];
 	interface I2D {
@@ -22,13 +27,9 @@
 		get: request.generator.post(endpoint),
 	};
 
+	let minDistUMAP: number;
+	let numNeighborsUMAP: number;
 	let initialProj: I2D[] = [];
-
-	const defaultParams = {
-		numNeighborsUMAP: 15,
-		minDistUMAP: 0.1,
-	};
-	let { minDistUMAP, numNeighborsUMAP } = defaultParams;
 
 	$: data = initialProj.map((d, i) => {
 		return {
@@ -51,6 +52,23 @@
 	async function project(config: IProject): Promise<I2D[]> {
 		const req = await mirror.post({ url: "project", payload: config });
 		return req;
+	}
+
+	/**
+	 * Get the column for id as an array of strings from
+	 * the arquero table.
+	 */
+	function curFilteredIds(): string[] {
+		return getIdsFromTable($filteredTable);
+	}
+	function getIdsFromTable(table: ColumnTable) {
+		const idColumnHash = columnHash($settings.idColumn);
+		const idColumnExists = table.size > 0 && table.column;
+		if (idColumnExists) {
+			return table.columnArray(idColumnHash) as string[];
+		} else {
+			return [];
+		}
 	}
 </script>
 
@@ -76,31 +94,5 @@
 		Visualize
 	</Button>
 
-	<!-- UMAP parameter customization for later -->
-	<div>
-		<h3>UMAP Parameters</h3>
-		<div>
-			<label for="min-dist">Min Dist</label>
-			<input
-				id="min-dist"
-				bind:value={minDistUMAP}
-				type="range"
-				step="0.1"
-				min="0.1"
-				max="3.0" />
-			{minDistUMAP}
-		</div>
-
-		<div>
-			<label for="nn">Num Neighbors</label>
-			<input
-				id="nn"
-				type="range"
-				bind:value={numNeighborsUMAP}
-				step="10"
-				min="5"
-				max="100" />
-			{numNeighborsUMAP}
-		</div>
-	</div>
+	<UMAPParams bind:minDistUMAP bind:numNeighborsUMAP />
 </div>
